+++ lightsim2grid/_aux_add_gen.py
@@ -5,7 +5,7 @@
 # you can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of LightSim2grid, LightSim2grid implements a c++ backend targeting the Grid2Op platform.
-
+import numpy as np
 
 def _aux_add_gen(model, pp_net):
     """
@@ -20,6 +20,10 @@ def _aux_add_gen(model, pp_net):
     -------
 
     """
+    if "parallel" in pp_net.gen and np.any(pp_net.gen["parallel"].values != 1):
+        raise RuntimeError("Cannot handle 'parallel' gen columns. Please duplicate the rows if that is the case. "
+                           "Some pp_net.line[\"parallel\"] != 1 it is not handled by lightsim yet.")
+
     model.init_generators(pp_net.gen["p_mw"].values,
                           pp_net.gen["vm_pu"].values,
                           pp_net.gen["min_q_mvar"].values,