+++ lightsim2grid/_aux_check_legit.py
@@ -50,3 +50,10 @@ def _aux_check_legit(pp_net):
     if "storage" in pp_net and pp_net.storage.shape[0]:
         raise RuntimeError("Unsupported element found (Storage - \"pp_net.storage\") "
                            "in pandapower network")
+
+    if pp_net.sn_mva != 1.:
+        raise RuntimeError("Pandapower network with sn_mva != 1 are not supported yet, especially in the "
+                           "conversion for some trafo (and probably somewhere else too)")
+    if pp_net["_options"]["trafo_model"] != "t":
+        raise RuntimeError("Cannot handle pp_net with trafo not modeled in \"t\". Please change "
+                           "your network and load it again.")